tests: Split LB test
