+++ testsuite/python/CMakeLists.txt
@@ -144,6 +144,7 @@ python_test(FILE integrator_npt.py MAX_NUM_PROC 4)
 python_test(FILE integrator_steepest_descent.py MAX_NUM_PROC 4)
 python_test(FILE dipolar_mdlc_p3m_scafacos_p2nfft.py MAX_NUM_PROC 1 LABELS long)
 python_test(FILE lb.py MAX_NUM_PROC 2 LABELS gpu)
+python_test(FILE lb_stats.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE force_cap.py MAX_NUM_PROC 2)
 python_test(FILE dpd.py MAX_NUM_PROC 4)
 python_test(FILE dpd_stats.py MAX_NUM_PROC 4)
